sidebar menu should not link to duplicate content [Youtrack AD-118]
The first link is a duplicate of the parent in content, but not in url - usability & SEO issue.
We skip inserting this dupe link.
(Springshare should be supplying a canonical meta, but that won't affect this menu issue)
Note the link can occur multiple times, because Springshare OTB shows links to the h2s down the page! :(